Move S3ConfigurationUpdater to S3 package
