+++ presto-hive/src/main/java/com/facebook/presto/hive/s3/PrestoS3ConfigurationUpdater.java
@@ -13,7 +13,6 @@
  */
 package com.facebook.presto.hive.s3;
 
-import com.facebook.presto.hive.S3ConfigurationUpdater;
 import io.airlift.units.DataSize;
 import io.airlift.units.Duration;
 import org.apache.hadoop.conf.Configuration;
